Merge branch 'upstream-1.6'

Conflicts:

	dspace-xmlui/dspace-xmlui-webapp/src/main/webapp/i18n/messages.xml
	dspace-xmlui/dspace-xmlui-webapp/src/main/webapp/themes/Reference/lib/style.css
	dspace-xmlui/dspace-xmlui-webapp/src/main/webapp/themes/Reference/sitemap.xmap
	dspace-xmlui/dspace-xmlui-webapp/src/main/webapp/themes/dri2xhtml/structural.xsl
	dspace/config/oaicat.properties
	dspace/config/xmlui.xconf

+++ dspace-api/src/main/java/org/dspace/checker/BitstreamInfo.java
@@ -1,6 +1,5 @@
 /*
- * Copyright (c) 2004-2005, Hewlett-Packard Company and Massachusetts
- * Institute of Technology.  All rights reserved.
+ * Copyright (c) 2002-2009, The DSpace Foundation.  All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are
@@ -13,8 +12,7 @@
  * notice, this list of conditions and the following disclaimer in the
  * documentation and/or other materials provided with the distribution.
  *
- * - Neither the name of the Hewlett-Packard Company nor the name of the
- * Massachusetts Institute of Technology nor the names of their
+ * - Neither the name of the DSpace Foundation nor the names of its
  * contributors may be used to endorse or promote products derived from
  * this software without specific prior written permission.
  *
@@ -141,7 +139,7 @@ public final class BitstreamInfo
      * @param procStartDate
      *            When the last bitstream check started.
      */
-    public BitstreamInfo(boolean del, int storeNo, int sz, String bitstrmFmt,
+    public BitstreamInfo(boolean del, int storeNo, long sz, String bitstrmFmt,
             int bitstrmId, String usrFmtDesc, String intrnlId, String src,
             String chksumAlgorthm, String chksum, String nm, Date procEndDate,
             boolean toBeProc, Date procStartDate)
@@ -204,7 +202,7 @@ public final class BitstreamInfo
      * 
      * @return int
      */
-    public int getSize()
+    public long getSize()
     {
         return dspaceBitstream.getSize();
     }
